通过websocket,在连接状态变化和节点信息变化时,相应页面通知刷新

+++ src/main/java/cn/zk/manager/AbstractCuratorManager.java
@@ -3,13 +3,14 @@ package cn.zk.manager;
 import cn.zk.app.config.CuratorManagerProperties;
 import cn.zk.common.AdminException;
 import cn.zk.common.RespCode;
-import cn.zk.manager.observer.ConnStateObserverDTO;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.curator.RetryPolicy;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
-import org.apache.curator.framework.state.ConnectionState;
+import org.apache.curator.framework.recipes.cache.PathChildrenCacheListener;
+import org.apache.curator.framework.recipes.cache.TreeCache;
+import org.apache.curator.framework.recipes.cache.TreeCacheListener;
 import org.apache.curator.framework.state.ConnectionStateListener;
 import org.apache.curator.retry.RetryNTimes;
 import org.apache.zookeeper.data.ACL;
@@ -27,7 +28,7 @@ import java.util.Observable;
  * @author zhubenle
  */
 @Slf4j
-public abstract class AbstractCuratorManager extends Observable implements ConnectionStateListener {
+public abstract class AbstractCuratorManager extends Observable implements ConnectionStateListener, PathChildrenCacheListener, TreeCacheListener {
 
     protected final static String DEFAULT_CHARSET = "UTF-8";
 
@@ -41,6 +42,18 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
                 .connectionTimeoutMs(properties.getConnectionTimeoutMs()).retryPolicy(retryPolicy).build();
         client.getConnectionStateListenable().addListener(this);
         client.start();
+        initCache();
+    }
+
+    @SneakyThrows
+    protected void initCache() {
+        TreeCache treeCache = new TreeCache(client, "/");
+        treeCache.getListenable().addListener(this);
+        treeCache.start();
+
+//        PathChildrenCache pathChildrenCache = new PathChildrenCache(client, "/", true);
+//        pathChildrenCache.getListenable().addListener(this);
+//        pathChildrenCache.start(PathChildrenCache.StartMode.NORMAL);
     }
 
     public void close() {
@@ -52,19 +65,8 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
         deleteObservers();
     }
 
-    @Override
-    public void stateChanged(CuratorFramework client, ConnectionState newState) {
-        setChanged();
-        notifyObservers(new ConnStateObserverDTO(newState, client.getZookeeperClient().getCurrentConnectionString()));
-        if (ConnectionState.LOST.equals(newState)) {
-            close();
-        }
-    }
-
-    public void checkState() {
-        if (!client.getZookeeperClient().isConnected()) {
-            throw new AdminException(RespCode.ERROR_10005);
-        }
+    public boolean isConnected() {
+        return client.getZookeeperClient().isConnected();
     }
 
     /**
@@ -78,21 +80,27 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
     @SneakyThrows
     public List<String> listChildrenPath(String parentPath) {
         log.debug("获取路径{}子路径名称列表", parentPath);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.getChildren().forPath(parentPath);
     }
 
     @SneakyThrows
     public void deletePath(String path, Integer version) {
         log.debug("删除路径{}, version={}", path, version);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         client.delete().guaranteed().deletingChildrenIfNeeded().withVersion(version).forPath(path);
     }
 
     @SneakyThrows
     public String createPath(String path, String data, List<ACL> acls, int createMode) {
         log.debug("创建路径{}, data={}, createMode={}", path, data, createMode);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.create().creatingParentsIfNeeded().forPath(path, data.getBytes());
     }
 
@@ -107,21 +115,27 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
     @SneakyThrows
     public boolean checkPathExist(String path) {
         log.debug("检查路径{}是否存在", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return Objects.nonNull(getPathStat(path));
     }
 
     @SneakyThrows
     public Stat getPathStat(String path) {
         log.debug("获取路径{}的元信息", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.checkExists().forPath(path);
     }
 
     @SneakyThrows
     public String getPathData(String path) {
         log.debug("获取路径{}的数据", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         byte[] data = client.getData().forPath(path);
         return Objects.nonNull(data) ? new String(data) : "";
     }
@@ -129,7 +143,9 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
     @SneakyThrows
     public String getPathData(String path, Stat stat) {
         log.debug("获取路径{}的数据", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         byte[] data = client.getData().storingStatIn(stat).forPath(path);
         return Objects.nonNull(data) ? new String(data) : "";
     }
@@ -137,21 +153,27 @@ public abstract class AbstractCuratorManager extends Observable implements Conne
     @SneakyThrows
     public Stat setPathData(String path, String data, Integer version) {
         log.debug("设置路径{}的数据", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.setData().withVersion(version).forPath(path, data.getBytes());
     }
 
     @SneakyThrows
     public Stat setACLs(String path, List<ACL> acls, Integer version) {
         log.debug("设置路径{}的ACL", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.setACL().withACL(acls).forPath(path);
     }
 
     @SneakyThrows
     public List<ACL> getACLs(String path) {
         log.debug("获取路径{}的ACL", path);
-        checkState();
+        if (!isConnected()) {
+            throw new AdminException(RespCode.ERROR_10005);
+        }
         return client.getACL().forPath(path);
     }
 }
